fix: perfil voltava vazio por excecao na sequencia
get_streak lancava "operator does not exist: date - bigint": ROW_NUMBER()
devolve bigint e o Postgres so tem operador `date - integer`. Faltava o cast.

O estrago era desproporcional. getProfile monta os dados com Future.wait, que
rejeita inteiro quando qualquer futuro lanca — um numero no card de sequencia
apagava nome, metas, pontos e bioimpedancia junto.

Alem do cast, as cinco chamadas secundarias passam a degradar individualmente.
So `profiles` e essencial; se o streak falhar de novo, aparece 0 e o resto da
tela continua de pe.

Cenarios verificados com usuario descartavel: sem treino (0), so hoje (1), tres
dias seguidos (3), buraco em ontem (1), ancora em ontem sem treino hoje (2),
ultimo treino ha tres dias (0), dois treinos no mesmo dia (1).

+++ supabase/migrations/20260818_fuso_local_e_streak.sql
@@ -78,7 +78,11 @@ BEGIN
             FROM workout_completions
            WHERE user_id = v_uid AND completed_date <= v_ancora) d
   ) t
-  WHERE completed_date = v_ancora - (rn - 1);
+  -- O cast para int é obrigatório: ROW_NUMBER() devolve bigint e o Postgres
+  -- só tem operador `date - integer`. Sem ele a função lança
+  -- "operator does not exist: date - bigint" e, como o perfil monta os dados
+  -- com Future.wait, a tela INTEIRA cai junto.
+  WHERE completed_date = v_ancora - (rn - 1)::int;
 
   RETURN COALESCE(v_total, 0);
 END;
